#93 Fixed Packaging in Tests after Refactoring
