+++ src/TestUtilities/test-utils.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { BrowserRouter as Router } from "react-router-dom";
-import {AuthenticationContext} from '../Authenticated';
+import {AuthenticationContext} from '../Firebase/Authenticated';
