+++ src/__tests__/personalProfileTest.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import PersonalProfilePage from '../Pages/PersonalProfile';
-import ImageGrid from '../Components/ImageGrid';
-import { act, render, screen,fireEvent} from "../test-utils";
+import ImageGrid from '../Components/Grids/ImageGrid';
+import { act, render, screen,fireEvent} from "../TestUtilities/test-utils";
 import { getByText } from '@testing-library/react';
-import {database} from '../firebase';
+import {database} from '../Firebase/firebase';
+import '@testing-library/jest-dom';
 
 
 beforeEach(async () => {
